+++ FE/src/components/Filter/Date/Date.jsx
@@ -1,50 +1,23 @@
 import React, { useState, useReducer } from "react";
 import styled from "styled-components";
-
-import "moment/locale/ko";
-import "react-dates/initialize";
-import "react-dates/lib/css/_datepicker.css";
-import { DayPickerRangeController } from "react-dates";
-import { START_DATE } from "react-dates/constants";
-
 import dateReducer from "Reducers/dateReducer";
 import { dateInitialState } from "InitialStates/initialStates";
-import { setDate, resetDate } from "Actions/dateAction";
+import { formatDate, toMonthDayString } from "Utils/utils";
 import FilterButton from "../FilterButton";
-import Modal from "../Modal";
+import DateModal from "./DateModal";
 
-const Date = () => {
+const Date = ({ dispatchHandler }) => {
   const [toggle, setToggle] = useState(false);
   const [dateState, dateDispatch] = useReducer(dateReducer, dateInitialState);
   const { startDate, endDate } = dateState;
-  const [focus, setFocus] = useState(START_DATE);
 
   const setFilterState = () => {
+    if (toggle)
+      dispatchHandler({
+        checkin: formatDate(dateState.startDate),
+        checkout: formatDate(dateState.endDate),
+      });
     setToggle(!toggle);
-  };
-
-  const onDatesChange = date => dateDispatch(setDate(date));
-
-  const onFocusChange = focusedInput => setFocus(!focusedInput ? START_DATE : focusedInput);
-
-  const modalContent = (
-    <DayPickerRangeController
-      numberOfMonths={2}
-      startDate={startDate}
-      endDate={endDate}
-      onDatesChange={onDatesChange}
-      focusedInput={focus}
-      onFocusChange={onFocusChange}
-    />
-  );
-
-  const modalOptions = {
-    contents: modalContent,
-    toggleHandler: () => setToggle(false),
-  };
-
-  const toMonthDayString = str => {
-    return `${str.month() + 1}월 ${str.date()}일`;
   };
 
   const renderDateButtonText = () => {
@@ -56,7 +29,7 @@ const Date = () => {
   return (
     <DateWrapper>
       <FilterButton clickHandler={setFilterState} active={toggle} text={renderDateButtonText()} />
-      {toggle && <Modal options={modalOptions} />}
+      {toggle && <DateModal setToggle={setFilterState} dateState={dateState} dateDispatch={dateDispatch} />}
     </DateWrapper>
   );
 };
